+++ designer/listener-api/src/main/scala/pl/touk/nussknacker/ui/listener/ProcessChangeEvent.scala
@@ -1,5 +1,6 @@
 package pl.touk.nussknacker.ui.listener
 
+import pl.touk.nussknacker.engine.api.Comment
 import pl.touk.nussknacker.engine.api.deployment.ProcessActionId
 import pl.touk.nussknacker.engine.api.deployment.ScenarioActionName
 import pl.touk.nussknacker.engine.api.process.{ProcessId, ProcessName, VersionId}