+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/ProcessService.scala
@@ -7,6 +7,7 @@ import cats.syntax.functor._
 import com.typesafe.scalalogging.LazyLogging
 import db.util.DBIOActionInstances.DB
 import io.circe.generic.JsonCodec
+import pl.touk.nussknacker.engine.api.Comment
 import pl.touk.nussknacker.engine.api.ProcessVersion
 import pl.touk.nussknacker.engine.api.component.ProcessingMode
 import pl.touk.nussknacker.engine.api.deployment.{DataFreshnessPolicy, ProcessAction, ScenarioActionName}
@@ -31,12 +32,7 @@ import pl.touk.nussknacker.ui.process.repository.ProcessDBQueryRepository.{
   ProcessNotFoundError,
   ProcessVersionNotFoundError
 }
-import pl.touk.nussknacker.ui.process.repository.ProcessRepository.{
-  CreateProcessAction,
-  ProcessCreated,
-  RemoteUserName,
-  UpdateProcessAction
-}
+import pl.touk.nussknacker.ui.process.repository.ProcessRepository._
 import pl.touk.nussknacker.ui.process.repository._
 import pl.touk.nussknacker.ui.security.api.LoggedUser
 import pl.touk.nussknacker.ui.uiresolving.UIProcessResolver
@@ -60,9 +56,18 @@ object ProcessService {
 
   @JsonCodec final case class UpdateScenarioCommand(
       scenarioGraph: ScenarioGraph,
-      comment: Option[UpdateProcessComment],
+      comment: Option[Comment],
       scenarioLabels: Option[List[String]],
       forwardedUserName: Option[RemoteUserName]
+  )
+
+  final case class MigrateScenarioCommand(
+      scenarioGraph: ScenarioGraph,
+      scenarioLabels: Option[List[String]],
+      forwardedUserName: Option[RemoteUserName],
+      sourceEnvironment: String,
+      targetEnvironment: String,
+      sourceScenarioVersionId: Option[VersionId],
   )
 
   object GetScenarioWithDetailsOptions {
@@ -139,6 +144,10 @@ trait ProcessService {
   ): Future[Validated[NuDesignerError, ProcessResponse]]
 
   def updateProcess(processIdWithName: ProcessIdWithName, action: UpdateScenarioCommand)(
+      implicit user: LoggedUser
+  ): Future[UpdateProcessResponse]
+
+  def migrateProcess(processIdWithName: ProcessIdWithName, action: MigrateScenarioCommand)(
       implicit user: LoggedUser
   ): Future[UpdateProcessResponse]
 
@@ -429,6 +438,45 @@ class DBProcessService(
           )
         }
     }
+
+  override def migrateProcess(processIdWithName: ProcessIdWithName, action: MigrateScenarioCommand)(
+      implicit user: LoggedUser
+  ): Future[UpdateProcessResponse] = {
+    withNotArchivedProcess(processIdWithName, "Can't migrate graph archived scenario.") { details =>
+      val processResolver = processResolverByProcessingType.forProcessingTypeUnsafe(details.processingType)
+      val scenarioLabels  = action.scenarioLabels.getOrElse(List.empty).map(ScenarioLabel.apply)
+      val validation =
+        FatalValidationError.saveNotAllowedAsError(
+          processResolver.validateBeforeUiResolving(
+            action.scenarioGraph,
+            details.name,
+            details.isFragment,
+            scenarioLabels
+          )
+        )
+      val substituted = processResolver.resolveExpressions(action.scenarioGraph, details.name, validation.typingInfo)
+      val migrateProcessAction = MigrateProcessAction(
+        processId = processIdWithName.id,
+        canonicalProcess = substituted,
+        labels = scenarioLabels,
+        increaseVersionWhenJsonNotChanged = false,
+        forwardedUserName = action.forwardedUserName,
+        sourceEnvironment = action.sourceEnvironment,
+        targetEnvironment = action.targetEnvironment,
+        sourceScenarioVersionId = action.sourceScenarioVersionId,
+      )
+      dbioRunner
+        .runInTransaction(processRepository.migrateProcess(migrateProcessAction))
+        .map { processUpdated =>
+          UpdateProcessResponse(
+            processUpdated.newVersion
+              .map(ProcessCreated(processIdWithName.id, _))
+              .map(toProcessResponse(processIdWithName.name, _)),
+            validation
+          )
+        }
+    }
+  }
 
   def importProcess(processId: ProcessIdWithName, jsonString: String)(
       implicit user: LoggedUser