+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/repository/ProcessRepository.scala
@@ -1,28 +1,23 @@
 package pl.touk.nussknacker.ui.process.repository
 
 import akka.http.scaladsl.model.HttpHeader
-import cats.implicits.toTraverseOps
 import com.typesafe.scalalogging.LazyLogging
 import db.util.DBIOActionInstances._
 import io.circe.generic.JsonCodec
-import pl.touk.nussknacker.engine.api.deployment.{ScenarioVersion, _}
+import pl.touk.nussknacker.engine.api.Comment
+import pl.touk.nussknacker.engine.api.deployment._
 import pl.touk.nussknacker.engine.api.process._
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
-import pl.touk.nussknacker.engine.migration.ProcessMigrations
+import pl.touk.nussknacker.engine.migration.{ProcessMigration, ProcessMigrations}
 import pl.touk.nussknacker.ui.db.entity.{ProcessEntityData, ProcessVersionEntityData}
 import pl.touk.nussknacker.ui.db.{DbRef, NuTables}
-import pl.touk.nussknacker.ui.listener.Comment
 import pl.touk.nussknacker.ui.process.label.ScenarioLabel
 import pl.touk.nussknacker.ui.process.processingtype.provider.ProcessingTypeDataProvider
 import pl.touk.nussknacker.ui.process.repository.ProcessDBQueryRepository._
-import pl.touk.nussknacker.ui.process.repository.ProcessRepository.{
-  CreateProcessAction,
-  ProcessCreated,
-  ProcessUpdated,
-  UpdateProcessAction
-}
+import pl.touk.nussknacker.ui.process.repository.ProcessRepository._
 import pl.touk.nussknacker.ui.process.repository.activities.ScenarioActivityRepository
 import pl.touk.nussknacker.ui.security.api.LoggedUser
+import pl.touk.nussknacker.ui.util.LoggedUserUtils.Ops
 import slick.dbio.DBIOAction
 
 import java.sql.Timestamp
@@ -68,16 +63,43 @@ object ProcessRepository {
       forwardedUserName: Option[RemoteUserName]
   )
 
+  sealed trait ModifyProcessAction {
+    protected val processId: ProcessId
+    val canonicalProcess: CanonicalProcess
+    val increaseVersionWhenJsonNotChanged: Boolean
+    val labels: List[ScenarioLabel]
+    val forwardedUserName: Option[RemoteUserName]
+    def id: ProcessIdWithName = ProcessIdWithName(processId, canonicalProcess.name)
+  }
+
   final case class UpdateProcessAction(
-      private val processId: ProcessId,
+      protected val processId: ProcessId,
       canonicalProcess: CanonicalProcess,
       comment: Option[Comment],
       labels: List[ScenarioLabel],
       increaseVersionWhenJsonNotChanged: Boolean,
       forwardedUserName: Option[RemoteUserName]
-  ) {
-    def id: ProcessIdWithName = ProcessIdWithName(processId, canonicalProcess.name)
-  }
+  ) extends ModifyProcessAction
+
+  final case class MigrateProcessAction(
+      protected val processId: ProcessId,
+      canonicalProcess: CanonicalProcess,
+      labels: List[ScenarioLabel],
+      increaseVersionWhenJsonNotChanged: Boolean,
+      forwardedUserName: Option[RemoteUserName],
+      sourceEnvironment: String,
+      targetEnvironment: String,
+      sourceScenarioVersionId: Option[VersionId],
+  ) extends ModifyProcessAction
+
+  final case class AutomaticProcessUpdateAction(
+      protected val processId: ProcessId,
+      canonicalProcess: CanonicalProcess,
+      labels: List[ScenarioLabel],
+      increaseVersionWhenJsonNotChanged: Boolean,
+      forwardedUserName: Option[RemoteUserName],
+      migrationsApplies: List[ProcessMigration]
+  ) extends ModifyProcessAction
 
   final case class ProcessUpdated(processId: ProcessId, oldVersion: Option[VersionId], newVersion: Option[VersionId])
 
@@ -89,6 +111,12 @@ trait ProcessRepository[F[_]] {
   def saveNewProcess(action: CreateProcessAction)(implicit loggedUser: LoggedUser): F[Option[ProcessCreated]]
 
   def updateProcess(action: UpdateProcessAction)(implicit loggedUser: LoggedUser): F[ProcessUpdated]
+
+  def migrateProcess(action: MigrateProcessAction)(implicit loggedUser: LoggedUser): F[ProcessUpdated]
+
+  def performAutomaticUpdate(
+      automaticProcessUpdateAction: AutomaticProcessUpdateAction,
+  )(implicit loggedUser: LoggedUser): F[ProcessUpdated]
 
   def archive(processId: ProcessIdWithName, isArchived: Boolean): F[Unit]
 
@@ -147,16 +175,24 @@ class DBProcessRepository(
         processesTable.filter(_.name === action.processName).result.headOption.flatMap {
           case Some(_) => DBIOAction.failed(ProcessAlreadyExists(action.processName.value))
           case None =>
-            (insertNew += processToSave)
-              .flatMap(entity =>
-                updateProcessInternal(
-                  ProcessIdWithName(entity.id, entity.name),
-                  action.canonicalProcess,
-                  increaseVersionWhenJsonNotChanged = false,
-                  userName = userName
+            for {
+              entity <- insertNew += processToSave
+              res <- updateProcessInternal(
+                ProcessIdWithName(entity.id, entity.name),
+                action.canonicalProcess,
+                increaseVersionWhenJsonNotChanged = false,
+                userName = userName
+              )
+              _ <- scenarioActivityRepository.addActivity(
+                ScenarioActivity.ScenarioCreated(
+                  scenarioId = ScenarioId(res.processId.value),
+                  scenarioActivityId = ScenarioActivityId.random,
+                  user = loggedUser.scenarioUser,
+                  date = clock.instant(),
+                  scenarioVersionId = res.newVersion.map(v => ScenarioVersionId(v.value))
                 )
               )
-              .map(res => res.newVersion.map(ProcessCreated(res.processId, _)))
+            } yield res.newVersion.map(ProcessCreated(res.processId, _))
         }
     }
   }
@@ -164,52 +200,98 @@ class DBProcessRepository(
   def updateProcess(
       updateProcessAction: UpdateProcessAction
   )(implicit loggedUser: LoggedUser): DB[ProcessUpdated] = {
-    val userName = updateProcessAction.forwardedUserName.map(_.name).getOrElse(loggedUser.username)
-
-    def addNewCommentToVersion(scenarioId: ProcessId, scenarioGraphVersionId: VersionId) = {
-      updateProcessAction.comment.map { comment =>
-        run(
-          scenarioActivityRepository.addActivity(
-            ScenarioActivity.ScenarioModified(
-              scenarioId = ScenarioId(scenarioId.value),
-              scenarioActivityId = ScenarioActivityId.random,
-              user = ScenarioUser(
-                id = Some(UserId(loggedUser.id)),
-                name = UserName(loggedUser.username),
-                impersonatedByUserId = loggedUser.impersonatingUserId.map(UserId.apply),
-                impersonatedByUserName = loggedUser.impersonatingUserName.map(UserName.apply)
-              ),
-              date = Instant.now(),
-              scenarioVersion = Some(ScenarioVersion(scenarioGraphVersionId.value)),
-              comment = ScenarioComment.Available(
-                comment = comment.value,
+    editProcess(
+      updateProcessAction,
+      (processId, versionId, _) =>
+        ScenarioActivity.ScenarioModified(
+          scenarioId = ScenarioId(processId.value),
+          scenarioActivityId = ScenarioActivityId.random,
+          user = loggedUser.scenarioUser,
+          date = Instant.now(),
+          scenarioVersionId = Some(ScenarioVersionId(versionId.value)),
+          comment = updateProcessAction.comment match {
+            case Some(comment) =>
+              ScenarioComment.Available(
+                comment = comment.content,
                 lastModifiedByUserName = UserName(loggedUser.username),
                 lastModifiedAt = clock.instant(),
               )
-            )
-          )
+            case None =>
+              ScenarioComment.Deleted(
+                deletedByUserName = UserName(loggedUser.username),
+                deletedAt = clock.instant(),
+              )
+          }
         )
-      }.sequence
+    )
+  }
+
+  def migrateProcess(
+      migrateProcessAction: MigrateProcessAction,
+  )(implicit loggedUser: LoggedUser): DB[ProcessUpdated] = {
+    editProcess(
+      migrateProcessAction,
+      (processId, versionId, user) =>
+        ScenarioActivity.IncomingMigration(
+          scenarioId = ScenarioId(processId.value),
+          scenarioActivityId = ScenarioActivityId.random,
+          user = loggedUser.scenarioUser,
+          date = clock.instant(),
+          scenarioVersionId = Some(ScenarioVersionId(versionId.value)),
+          sourceEnvironment = Environment(migrateProcessAction.sourceEnvironment),
+          sourceUser = UserName(user),
+          sourceScenarioVersionId = migrateProcessAction.sourceScenarioVersionId.map(v => ScenarioVersionId(v.value)),
+          targetEnvironment = Some(Environment(migrateProcessAction.targetEnvironment)),
+        )
+    )
+  }
+
+  def performAutomaticUpdate(
+      automaticProcessUpdateAction: AutomaticProcessUpdateAction,
+  )(implicit loggedUser: LoggedUser): DB[ProcessUpdated] = {
+    editProcess(
+      automaticProcessUpdateAction,
+      (processId, versionId, _) =>
+        ScenarioActivity.AutomaticUpdate(
+          scenarioId = ScenarioId(processId.value),
+          scenarioActivityId = ScenarioActivityId.random,
+          user = loggedUser.scenarioUser,
+          date = Instant.now(),
+          scenarioVersionId = Some(ScenarioVersionId(versionId.value)),
+          changes = automaticProcessUpdateAction.migrationsApplies.map(_.description).mkString(", "),
+          errorMessage = None,
+        )
+    )
+  }
+
+  def editProcess[ACTION <: ModifyProcessAction](
+      action: ACTION,
+      activityCreator: (ProcessId, VersionId, String) => ScenarioActivity,
+  )(implicit loggedUser: LoggedUser): DB[ProcessUpdated] = {
+    val userName = action.forwardedUserName.map(_.name).getOrElse(loggedUser.username)
+
+    def addScenarioModifiedActivity(scenarioId: ProcessId, scenarioGraphVersionId: VersionId) = {
+      run(scenarioActivityRepository.addActivity(activityCreator(scenarioId, scenarioGraphVersionId, userName)))
     }
 
     for {
       updateProcessRes <- updateProcessInternal(
-        updateProcessAction.id,
-        updateProcessAction.canonicalProcess,
-        updateProcessAction.increaseVersionWhenJsonNotChanged,
+        action.id,
+        action.canonicalProcess,
+        action.increaseVersionWhenJsonNotChanged,
         userName
       )
       _ <- updateProcessRes match {
         // Comment should be added via ProcessService not to mix this repository responsibility.
         case updateProcessRes @ ProcessUpdated(processId, _, Some(newVersion)) =>
-          addNewCommentToVersion(processId, newVersion).map(_ => updateProcessRes)
+          addScenarioModifiedActivity(processId, newVersion).map(_ => updateProcessRes)
         case updateProcessRes @ ProcessUpdated(processId, Some(oldVersion), _) =>
-          addNewCommentToVersion(processId, oldVersion).map(_ => updateProcessRes)
+          addScenarioModifiedActivity(processId, oldVersion).map(_ => updateProcessRes)
         case _ => dbMonad.unit
       }
       _ <- scenarioLabelsRepository.overwriteLabels(
-        updateProcessAction.id.id,
-        updateProcessAction.labels
+        action.id.id,
+        action.labels
       )
     } yield updateProcessRes
   }
@@ -299,10 +381,7 @@ class DBProcessRepository(
     val updateNameInProcess =
       processesTable.filter(_.id === process.id).map(_.name).update(newName)
 
-    // Comment relates to specific version (in this case last version). Last version could be extracted in one of the
-    // above queries, but for sake of readability we perform separate query for this matter
-    // TODO: remove this comment in favour of process-audit-log
-    val addCommentAction = processVersionsTableWithUnit
+    val addScenarioNameChangedActivity = processVersionsTableWithUnit
       .filter(_.processId === process.id)
       .sortBy(_.id.desc)
       .result
@@ -313,14 +392,9 @@ class DBProcessRepository(
             ScenarioActivity.ScenarioNameChanged(
               scenarioId = ScenarioId(process.id.value),
               scenarioActivityId = ScenarioActivityId.random,
-              user = ScenarioUser(
-                id = Some(UserId(loggedUser.id)),
-                name = UserName(loggedUser.username),
-                impersonatedByUserId = loggedUser.impersonatingUserId.map(UserId.apply),
-                impersonatedByUserName = loggedUser.impersonatingUserName.map(UserName.apply)
-              ),
+              user = loggedUser.scenarioUser,
               date = Instant.now(),
-              scenarioVersion = Some(ScenarioVersion(version.id.value)),
+              scenarioVersionId = Some(ScenarioVersionId(version.id.value)),
               oldName = process.name.value,
               newName = newName.value
             )
@@ -335,7 +409,7 @@ class DBProcessRepository(
           .seq[Effect.All](
             updateNameInProcess,
             updateNameInProcessJson,
-            addCommentAction
+            addScenarioNameChangedActivity
           )
           .map(_ => ())
           .transactionally