+++ docs/Changelog.md
@@ -56,9 +56,9 @@
 * Scenario activities mechanism replacing old process actions:
   * [#6822](https://github.com/TouK/nussknacker/pull/6822), [#6929](https://github.com/TouK/nussknacker/pull/6929)
     * Scenario Activity API contract (without BE implementation)
-  * [#6823](https://github.com/TouK/nussknacker/pull/6823) 
+  * [#6823](https://github.com/TouK/nussknacker/pull/6823) , [#6962](https://github.com/TouK/nussknacker/pull/6962)
     * old process actions and comments migrated to new table in the db
-    * Scenario Activity API implementation
+    * Scenario Activity API BE implementation
 * [#6925](https://github.com/TouK/nussknacker/pull/6925) Fix situation when preset labels were presented as `null` when node didn't pass the validation.
 * [#6935](https://github.com/TouK/nussknacker/pull/6935) Spel: Scenario labels added to meta variable - `#meta.scenarioLabels`
 * [#6952](https://github.com/TouK/nussknacker/pull/6952) Improvement: TypeInformation support for scala.Option